Suppress display of "edited" info for forum topics
As explained in ticket #1235, showing the "edited" time info when
showing an individual forum topic looks misleading because it makes
it look like the first post (the one which started the topic) was
itself edited, when all that really happened was that someone
replied to it and so the "updated at" field was changed.

We need to change the field so that topics are correctly sorted in
order, but we don't want to see the field when viewing an
individual topic.

We could add a second timestamp here to distinguish between "last
replied to" and "last edited", but given that topic text is so
seldom edited I don't think it's worth the complexity.

For more details see:

  https://wincent.com/issues/1235

Signed-off-by: Wincent Colaiuta <win@wincent.com>