Move "receive" from issue mailer into dedicated support mailer
Although we will initially just be handling incoming support
requests send to support@wincent.com, eventually this mail handler
will also take care of replies to forum topics and the like, so it
makes sense to move it into a dedicated handler of its own rather
than tying it to the issue mailer.

The name "support mailer" was chosen because it handles mail for
"support@wincent.com".

Signed-off-by: Wincent Colaiuta <win@wincent.com>