Create Message records for incoming mail
Currently don't do anything more than that, but at least now
we record incoming email and extract the plaintext body from
it. Email without any plaintext is probably spam, but we
sanitize it (strip out HTML tags, clean up the whitespace so
that it won't look too horrid when viewed as wikitext) and
record it anyway.

Signed-off-by: Wincent Colaiuta <win@wincent.com>