Workaround for Rails 2.3.0 RC1 bug #2043
Rails 2.3.0 RC1 changes the behaviour of "polymorphic_url" so
that it returns different formats depending on where it is
called from.

This means that "read more" links in Atom feeds, which previously
took you to the HTML representation of the resource, will now
try to take you to the Atom representation.

Most likely not what the user expects, and a method that changes
its behaviour depending on where it's being called is nasty in
any case.

See:

  http://rails.lighthouseapp.com/projects/8994/tickets/2043

Signed-off-by: Wincent Colaiuta <win@wincent.com>