Make nginx page caching play nicely with non-GET requests
Surprising that I hadn't noticed this before, but the "canonical"
way of doing page caching with nginx has a serious incompatibility
with certain request types.

  # cached pages
  if (-f $request_filename) {
    rewrite (.*) $1.html break;
  }

Let's take an example using this configuration:

  1. Client requests /twitter
  2. Index action runs, page cached to /twitter.html
  3. Client requests /twitter/new
  4. Submits form, POSTs to /twitter
  5. nginx sees /twitter.html but can't POST to it, obviously

So the result is a 405 "Not allowed" error and no new records
can be created. I'd never seen this problem before because until
now I had only page-cached Atom feeds.

The solution, then, is to only let nginx's page caching kick in
for GET requests, but it is not as easy as you might thinking
because the nginx rewrite module neither supports nested "if"
blocks nor "AND" or "&&" logic in "if" statements.

The hack implemented in this commit seems to be the least hacky
way of working around this shortcoming.

Signed-off-by: Wincent Colaiuta <win@wincent.com>