Turn on page caching even in test environment
This is the only way we can test page caching, and evidently
we do want to test it.

The problem is that rails turns "caches_page" into a total no-op
when page caching is disabled, and that means we can't even set
mock message expectations on things like "cache_page" (it's not
that the after filters do nothing; it's that the after filters
are not even set up).

So this looks like a necessary evil.

Signed-off-by: Wincent Colaiuta <win@wincent.com>