Use a more defensive format check in "require_admin"
Normally the "format" parameter for HTML requests will be
blank, but just in case it isn't, we were checking for an
explicit "html" as well.

To be even more defensive, use a case-insensitive regex so
that we'll also catch variants like "HTML" or "Html" if
they ever make it through.

Of course, either way the method is still secure because
it still prevents access to the requested resource: the
only difference is whether the denial results in a redirect
or in an error.

Signed-off-by: Wincent Colaiuta <win@wincent.com>