Remove sidebar from application layout
Given that the sidebar requires the non-sidebar part of the content
to have special CSS applied, and not all pages are going to want a
sidebar, remove it from the application layout.

Specific controllers/actions which will want to show a sidebar (the
posts#index action, for example) can insert a sidebar in their
specific template.

Signed-off-by: Wincent Colaiuta <win@wincent.com>