Add GNU Affero General Public License to application.js
"application.js" is starting to acquire more and more
functionality, and seeing as it is distributed in source
form it is probably time to think about what terms use of
the source should be permitted.

So release it under the GNU Affero General Public License
and include a copy of the license on the server.

Signed-off-by: Wincent Colaiuta <win@wincent.com>