Make flashes compatible with page caching
Send flashes to the browser using cookies, and let the browser
show them using JavaScript. The browser immediately deletes
the cookie so as to avoid seeing a "double flash".

In this way the flash itself never appears in the HTML source,
so when the source is page cached the flash does not get
frozen into it.

Downside: users without JavaScript enabled can't see flashes.

Impact: only a minimum of users surf without JavaScript, and
only a minimum of (infrequent) actions produces flashes, _and_
consequences of not seeing a flash are usually insignificant
seeing as most flashes are purely informative and about
non-critical events (such as "Logged in successfully"), so not
much impact.

Certainly seems a worthwhile trade off because of the payoffs
in making page-caching easier. This in turn makes more complex,
useful pages possible. The other workarounds, turning off
flashes entirely for cached pages, or not page-caching, are not
attractive.

Signed-off-by: Wincent Colaiuta <win@wincent.com>