Apply upstream XSS/CRSF security fixes
Here I adapt upstream commit 8c5487e, "Security fixes for XSS and
CSRF issues" to the local version of the plugin.

Note that some of these are formatting changes in order to bring
the local version closer to the upstream version and make it
easier to apply such fixes in the future.

Signed-off-by: Wincent Colaiuta <win@wincent.com>