Remove comment (upstream bug fixed)
The "reset_session" breakage is now fixed upstream,
so remove comments about it. Nevertheless, I am
leaving my workaround in place just in case, and
because it might save us from a trip to the
database in any case.

I still need to look and see if I can create a spec
to catch the original bug.

See:

  http://github.com/rails/rails/commit/9b1b88f09cf1498f04e1cd469d0d5ffccf2b93cc

Signed-off-by: Wincent Colaiuta <win@wincent.com>