New AJAX in-place edit for issue summaries
This is part of a work-in-progress. This, the first step, is
basically a hand-coded, unobtrusive in-place editor for one
field (the summary) on the Issues#show page.

The next step will be to refactor the code and make it usable
from multiple sites (ie. without hardcoded URLs and the like).

Heavily inspired by "REST in Place" by Jan Varwig, but with a
couple of discrepancies in the design which in turn lead to a
different implementation:

  - this should definitely not require non-standard HTML
    attributes which would cause page validation to fail
  - this should be truly unobtrusive, not requiring the
    insertion of _anything_ additional into the HTML
  - this should highlight on mouse hover to provide visual
    feedback that the area under the cursor is editable
  - it should activate on double-click, not single-click,
    to allow for comfortable text selection and avoid unintended
    editing
  - it should provide feedback in the face of errors, and allow
    re-editing and re-submission in such cases
  - it should avoid hitting the server with two AJAX requests
    for each edit, and instead just to one request and use the
    response from that to update the page

Signed-off-by: Wincent Colaiuta <win@wincent.com>