Add sanitization to new in-place editor implementation
Add an "escapeHTML" function so that even malicious text input
like:

  <script>alert("owned!");</script>

Can be safely handled in the new in-place editor. Note that
it is not necessary to use this function everywhere, because
in many places merely swapping in the JQuery "text()" function
for the "html()" function performs adequate escaping for us.

Signed-off-by: Wincent Colaiuta <win@wincent.com>