Fix inadequate HTML escaping in "link_to_commentable"
Was a security hole.

This problem was originally noticed on the user dashboard
page but it affects all sites where the "link_to_commentable"
helper method is used. As such, the only public-facing site
where the vulnerability was exposed was the dashboard itself;
the other affected site was the admin-only comments index.

Was a security hole: a malicious issue summary like
"<script>alert('foo');</script>" would be executed if the
issue was referenced in the "recent comments" listing.

Signed-off-by: Wincent Colaiuta <win@wincent.com>