Rename "configure_ajax_select" to "ajax_select"
Although "configure_ajax_select" might be a little closer to
what the actual purpose of the function is, "ajax_select" is
sufficiently unique for it to serve as an identifier.

This change brings the function into line with the corresponding
"in-place" editor function, which is just called "in_place_edit".

Signed-off-by: Wincent Colaiuta <win@wincent.com>