Make "acts as searchable" work with models without a "public" attribute
We want to make the tweet model searchable, but "acts as searchable"
requires models to have a "public" field. We therefore have two
options: either add the field to the tweet model, or modify "acts as
searchable" to work with models that don't have the field.

This commit implements the latter options because it requires the
least invasive changes to the code. Adding an unnecessary "public"
attribute to the tweet model would add complexity.

Signed-off-by: Wincent Colaiuta <win@wincent.com>