Show search field at top of search results pages
The search field is pre-populated with the query used to
produce those results.

This is a nice usability improvement for two reasons:
firstly, it provides feedback about exactly what query was
used to generate a given set of results; and secondly, it
makes it easy to further expand or refine a given search.

At the same time we change the "search again" button label
to "new search". This makes it a little clearer what the
purpose of the button is, and differentiates it from the
"search" button which is added to the top of each page by
this commit.

We probably still need to make the search tips a little
bit more discoverable. At the moment the only way you'll
see them is if you hit that "new search" button.

See:

  https://wincent.com/issues/1259

Signed-off-by: Wincent Colaiuta <win@wincent.com>