Remove visible_comments method from Issue model
Somewhere along the line Rails got smart enough to not
need a separate method here. It now respects the ":include"
specification on the association, so
"@issue.comments.published" is now basically equivalent to
"@issue.visible_comments".

According to my testing using Rails 2.3.2, both effectively
first pull in all the comments, and then as a separate
query slurp in all the users.

Signed-off-by: Wincent Colaiuta <win@wincent.com>