Add improved icons
New "external", "mailto", and pagination icons. Some of these
are copies of, and others are derived works based on, icons
by Yusuke Kamiyamane and licensed under a Creative Commons
Attribution License.

Seeing as we need to provide attribution, we remove the old
copyright notice from the footer and instead provide a link
to a separate "Copyright and legal information" page where
we can provide extensive attribution details for the icons
and everything else in the application as well.

See:
  http://www.pinvoke.com/

Signed-off-by: Wincent Colaiuta <win@wincent.com>