Cucumber: get "issue" features working again
- Use HTML label tags to help Webrat find login form
  fields
- Make feature briefer by creating verified email
  addresses by default in FixtureReplacement
- Likewise, create verified users and use a default
  passphrase in FixtureReplacement
- Update expected text to match actual text
- Add "cacheable_flash" helper to make it easy to
  check for successful login
- Tweak user spec to fix breakage caused by changes
  to FixtureReplacement defaults

With this change the spec and feature suites both run
successfully, but this is still very much a work in
progress.

I suspect, for example, that the handling of cacheable
flashes may need to be modified (does Webrat actually
delete the cookie like the browser does on displaying
the flash? I don't think it does, so flash messages
probably just accumulate as we move from page to page
with Webrat).

I am also not entirely happy with the circular
references in the FixtureReplacement object graph
(new_email creates a user, but new_user doesn't create
an email). This leaves us in the unpleasant situation
of having "create_user" yield a verified user with no
email address.

Signed-off-by: Wincent Colaiuta <win@wincent.com>