Remove "spam" attribute from all models which have it
The "spam" attribute has been in the codebase for a long
time now, and it turns out to have been a misguided and
unnecessary complication all this time.

Originally, the idea was that by keeping spam messages
around in the database it would be easier to do Bayesian
training.

In reality, Bayesian filtering has yet to be implemented,
in practice I have always just used the "Destroy" rather
than the "Spam" buttons, and models or their content for
any future Bayesian training implementation can easily
be kept in a separate database table.

The cost of maintaining this "spam" attribute is that it
has made many, many queries throughout the codebase more
complicated (more likely to have bugs, harder to code,
slower).

So, out it comes.

See:
  https://wincent.com/issues/1264
  https://wincent.com/issues/734

Signed-off-by: Wincent Colaiuta <win@wincent.com>