Revert "Capitalize article titles when not found"
This reverts commit 589a102. The original motivation of
that commit was to propose an article title like "Tagging"
when the user had followed a link in a sentence like "one
of the better methods is [[tagging]] the content".

While that might be a nice convenience, experience has
shown that it actually causes a lot of hassle when following
links like "Updating to Git 1.6.2.2 on Mac OS X Leopard
10.5.6", which would propose an article title like "Updating
to git 1.6.2.2 on mac os x leopard 10.5.6", seeing as the
capitalize method changes the first characer to uppercase
and everything else in the string to lowercase.

A better compromise might be to capitalize _only_ the first
word, but even that might be unwanted in the case of
links like "[[WOPublic]]" and "[[IBM]]" etc. Perhaps the
correct behaviour should be to capitalize the first word if
and only if it is all lowercase to begin with.

Signed-off-by: Wincent Colaiuta <win@wincent.com>