Skip request forgery protection in search controller
The static pages on the site don't have request forgery
protection tokens (anti-CSRF tokens) in their search forms
for obvious reasons (the tokens are dynamic and those
pages are static) and this in turn means that the search
fields are useless because any attempt to use them results
in a 422 "the change you want was rejected" error.

Seeing as search requests are "idempotent" (ie they always
produce the same result if repeated) and have no side
effects, it should be safe for the search controller to
"opt out" of the request forgery protection system, thus
making searches from static pages work again. This is of
particular interest in the case of the 404 (not found)
page, where a very logical user action would be to
execute a search.

Note that it even though user identity _is_ involved in
the search results (that is, admins will see more results
than non-admin users), the fact that searching has no
side effects makes CSRF requests to the search form
totally innocuous. (In order to be useful a CSRF attack
must have some kind of desired side effect, like changing
a login passphrase.)

See the Wikipedia article on CSRF for more information:

  http://en.wikipedia.org/wiki/Cross-site_request_forgery

Signed-off-by: Wincent Colaiuta <win@wincent.com>