Better nil handling in set_protected_attribute method
It is possible (although unlikely) that a param could
be explicitly set to nil, in which case we would fail
to take any action.

In reality this hasn't proven to be a problem so far
because explicitly deleted or "unset" values appear in
the params hash as empty strings rather than nil
values; even so, it's not a bad idea to be explicit
here: we are interested in knowing whether a given
key has had a value set for it, not whether said value
is "true" or not.

Signed-off-by: Wincent Colaiuta <win@wincent.com>