Use set_protected_attribute to set pending tags in issues controller
It may save some head-scratching later on if we always use the same
method for setting protected attributes (even virtual attributes like
this one which don't have a corresponding column in the database
table).

I am not yet convinced that the set_protected_attribute method is
actually the best API for this case, so I am not using it extensively
yet. In the case of the JS format in the "update" action, however, I
do think it's not a bad idea to use it for consistency with the other
protected attribute update that's taking place right next to this
one.

Signed-off-by: Wincent Colaiuta <win@wincent.com>