Make search field in nav bar use jQuery instead of Prototype
There was a strange glitch in the move here from Prototype to
jQuery: namely, in both Firefox and Safari, repeatedly toggling
the search field on and off would cause it to move to the left.

The workaround was to hide the text field only rather than the
entire form.

Note that unlike the issue search field in the previous commit,
this particularly field already degrades gracefully when no JS
is available.

Signed-off-by: Wincent Colaiuta <win@wincent.com>