Partial replacement of in-place editing for links#index
This replaces the existing Prototype/Scriptaculous in-place-editing
implementation with the unobtrusive jQuery one originally used in
the issues#show action.

The current implementation acts only on the first link
in the list; it needs to be extended to match all elements of
the specified class and install in-place-editor fields for all
of them.

Signed-off-by: Wincent Colaiuta <win@wincent.com>