Highlight errors using red CSS after an AJAX error
The in-place editor field current preserves the new value
in the face of an AJAX error, but it doesn't really provide
much feedback to the user other than the inital JS alert
that appears.

We want to preserve the new text so that the user can easily
try editing it to correct the error and then resubmit. But
the problem with conserving the text is that the user may
mistakenly think that it really was saved.

We get around this by changing the background colour to a
light red (really a salmon color) so that the user will
see exactly where the problem was if an AJAX request fails,
and is unlikely to mistakenly believe that the value currently
reflected in the view is what's been saved to the database.

Signed-off-by: Wincent Colaiuta <win@wincent.com>