Don't check authenticity tokens for any AJAX requests
In theory AJAX requests aren't a possible vehicle for CRSF attacks,
so checking them for valid tokens is an unnecessary complication.
(In the standard CRSF attack the user must click on a malicious link
which will then trigger some state-altering action on the server
where the victim has an account or an active session. With current
browsers such a malicious link can never be an AJAX request because
browsers follow a "same origin" rule that will only allow AJAX
requests to go back to the originating server, not some other
server. If the attacker manages to implant a malicious link on the
same originating server, then the CRSF attack is kind of moot
anyway, because by that time they already have the ability to
execute any AJAX requests they want at all.)

This is a backport of commit 256b0ee from the Rails master branch.
I decided to backport it because local commit 6bb575b, which caches
the forgery protection token from the global application layout, is
going to wreak havoc with page caching, seeing as the token will be
inappropriately cached on some pages.

Signed-off-by: Wincent Colaiuta <win@wincent.com>