jQuery version of live preview in articles#new and articles#edit
This is just a straight replacement of the original implementation
(no attempt has been made yet to turn this into Unobtrusive JS).

Evidently can and should be refactored in the future to reduce
duplication, but I will need to start converting the other
controller and actions over before I can tell exactly what will
be needed in order to achieve this.

Also still left to do is to refrain from submitting if the
field contents haven't actually changed.

Signed-off-by: Wincent Colaiuta <win@wincent.com>