Fix "misc" controller breakage in admin namespace
With the removal of the default "catch-all" routes the links to the
"misc" controller (such as the "legal info" page at the bottom of the
application layout) were broken when in the admin namespace.

This is because the url_for method tries to link to "admin/misc/legal"
instead of "misc/legal", and a routing error gets raised.

One possible workaround is to just use a hard-coded string in the
layout ("/misc/legal"), but for now have gone with just adding the
equivalent route in the admin namespace (ie. "/admin/misc/legal" and
so on now work).

Signed-off-by: Wincent Colaiuta <win@wincent.com>