Use hard-coded links to work around url_for quirks (test env)
While url_for() works fine in development and production, it
is broken for links in the "misc" controller in many many
tests.

I suspect this is due to anomalies in the test environment (the
app insists on SSL and the best I can do is fake it). During
view specs "no route matches" errors are raised for url_for
calls with "controller => misc, action => legal", for example.
(Note that the protocol parameter is suppressed.)

One workaround is to set up a new route without the protocol,
but that feels just wrong.

Another workaround is to just forget about doing any specs
for views which call url_for; again not a very realistic option.

In the end we take the lesser of all evils approach and instead
of using "url_for" just hardcode string links to "/misc/legal"
and the like.

Signed-off-by: Wincent Colaiuta <win@wincent.com>