Drop "fieldName" parameter to setup_preview_link
In the case of field observers it makes sense to have a principal
field that is being observed, identified by "fieldName", and an
optional array of related fields which should be included in the
preview, identified by "included".

But in the case of the setup_preview_link, there is no "principal"
field that we are interested in; we are only interested in the
preview of all the relevant fields as a whole.

So refactor the setup_preview_link function to always operate on an
array of "included" fields and not use the "fieldName" parameter at
all.

Signed-off-by: Wincent Colaiuta <win@wincent.com>