Make authentication return non-empty response text on access errors
Return "forbidden" rather than an empty response. It is unlikely that this
error will come up in a normal error request; it possibly could if a user
leaves a page open for a long time or in some other way expires his or her
session and then initiates the AJAX request.

But even though it is unlikely, it is nice to provide some informative
error text just in case (now that the AJAX functions are being upgraded
to display flashes on errors).

Signed-off-by: Wincent Colaiuta <win@wincent.com>