Fix specs broken by recent changes
I've been accumulating a bit of "technical debt" lately in order
to move development forward as quickly as possible. So that means
sometimes rushing forward with a series of commits and only later
coming back and updating the spec suite. (I'm doing this because
the suite takes about a minute to run.)

Signed-off-by: Wincent Colaiuta <win@wincent.com>