Make selection matching in ajax_select case-insensitive
When trying to find the current selection in ajax_select we were
calling toLowerCase() on the static text (so for example, "Feature
request" would become "feature request") and this matched correctly
against the supplied options hash (produced with the
"underscores_to_spaces" helper method).

But this didn't work for the products pop-up because the options
hash used mixed-case names like "WOCommon" and so forth.

There are three ways to fix this:

- stop calling "humanize" in the issues#show view, thus keeping
  "feature request" as is rather than turning it into "Feature
  request"; this in turn would allow us to drop the toLowerCase()
  call entirely
- do case-insensitive matching (ie. call toLowerCase on both sides)
- use "humanized" text in the pop-ups as well

Of these the second solution is the one I went with in the
interests of a quick hot-fix. But in the longer term (for the next
deployment) I am going to go with the first solution. The third
solution was never an attractive one as such text looks ugly and
I don't like the inconsitency between different pop-ups (status
and kind being "humanized" but product being left alone).

Fixes:

  https://wincent.com/issues/1288

Signed-off-by: Wincent Colaiuta <win@wincent.com>