Use "_path" instead of "_url" helpers where possible
The decision to use "_uri" helpers instead of "_path" ones to
help with the enforcement of SSL-only acccess to the site has
created some headaches:

- hacky patches to the "ssl?" method in both spec/spec_helper.rb
  and features/support/env.rb for RSpec and Cucumber respectively
- problems trying to get Selenium to work (given that it doesn't
  cope well with SSL-only forced redirects)

It turns out that enforcing at the application level was never a
good idea; it is a problem much better handled outside the app:
through firewall configuration (or simply instructing mongrel to
only listen on the loopback interface) and proper nginx setup.

We already had the right nginx setup in place (HTTP requests
are redirected to the HTTPS equivalent), and now we also have
adequate Mongrel configuration (firewall still to come).

So we are now free to drop all the SSL enforcement stuff and
switch back to using "_path" rather than "_url" helpers. There
a couple of benefits from this:

- no more need for kludges
- we can drop a before filter
- code simplification

This last benefit is only just manifesting itself with this
commit (in the reduction in lines of code), but further gains
should be had when we start dropping explicit ":protocol => 'https'"
from the code.

For further discussion see:

  https://wincent.com/issues/1131

Signed-off-by: Wincent Colaiuta <win@wincent.com>