Don't use "secure" cookies in local testing
We want to be able to test the application if need be by
connecting directly to the Mongrel instance via HTTP rather
than using HTTPS and the nginx front-end.

But there is a problem: you can login but the "Secure" login
cookie will never be sent to the Mongrel instance over an
HTTP connection.

So, for local testing only, we use "insecure" cookies. We
do this in the hope that we can get Selenium tests working
without any SSL certificate hassles or the need for complicated
and brittle Firefox workarounds.

Signed-off-by: Wincent Colaiuta <win@wincent.com>