Add "selenium" environment for Webrat
According to the Webrat docs, this environment is required when
running in "selenium" mode. So this is literally just a copy of
the test environment for now.

We also specify the port number where we want our Mongrel running,
because if Selenium tries to talk to the nginx front-end instead
of the Mongrel instance (and that is exactly what it appears to
want to do) then we had better make sure that nginx knows where
to find the instance.

Signed-off-by: Wincent Colaiuta <win@wincent.com>