Make find_recent methods more consistent
Make sure all models which implement a "find_recent" method
employ the same pattern (accepting an optional options hash
rather than an optional paginator).

This is an improvement not only in the name of consistency,
but also in the name of better separation between the model
and controller layers (the model layer should never really
have to know about the existence of such a thing as a
paginator).

Signed-off-by: Wincent Colaiuta <win@wincent.com>