Rename "_preview.js.haml" templates to "_preview.html.haml"
This makes the article, post and topic preview templates consistent
with the tweet preview template.

The problem here is caused because I am abusing the "js" format
of a RESTful controller action in order to send HTML previews. This
allows me to render standard HTML if the respond_to block specifies
"html", and render a live AJAX preview if it specifies "js".

But in both cases I'm really sending HTML, and that's why it's
really an abuse. (Nevertheless, it works pretty well). The alternative
would be to perform an explicit "request.xhr?" test inside the "html"
format block instead, and that is probably what I will eventualyl do.

In the meantime, however, this works well, but things needed to be
made consistent.

The old name allowed us to just render "_preview" and Rails would
automatically choose the right template, but it was very misleading
because the template didn't produce JavaScript at all.

The new name requires us to explicitly render "_preview.html.haml",
but now at least the template name actually reflects what will be
rendered.

The other benefit of renaming those template files is that when I
rewrite those controllers to stop the "abuse", the templates will
already have the right names, and if I want I can at that point
drop the explicit ".html.haml" extensions.

Fixes:
  https://wincent.com/issues/1239

Signed-off-by: Wincent Colaiuta <win@wincent.com>