Stop abusing format.js for live previews
Previously we were abusing "respond_to" with "format.js" to
provide live AJAX previews from inside a RESTful controller
action (the update action) in various controllers.

This was an abuse because while we were requesting a
JavaScript response, what we actually wanted was HTML.

Remove the pattern entirely, replacing it with a specific
check for "request.xhr?".

Signed-off-by: Wincent Colaiuta <win@wincent.com>