Make posts#show tag display match that of articles#show
Previously, if there were no tags we just omitted the display
entirely, showing a "Tags" heading with nothing underneath it.

Now we match the behaviour of articles#show by rendering the
"shared/taggings" partial, which shows the text "none" in the
event that there are no tags.

This has the added benefit of making things look more consistent
in terms of styling (because we now use an "h2" instead of an "h1"
heading for "Tags" for both actions), and probably more appropriate
anyway (the old heading looked too important relative to the "h2"
comments heading that could appear immediately under it).

Closes:

  https://wincent.com/issues/1211

Signed-off-by: Wincent Colaiuta <win@wincent.com>