Capistrano: check for "mkdtemp" gem
The mkdtemp gem is required in the spec suite (whenever
"rake spec" is run) but we don't really want to freeze
it into the "vendor" directory as it is another native
extension requiring compilation, and we really only
want to use it at test time.

So for now, just make sure that it's available before
deploying using Capistrano.

Signed-off-by: Wincent Colaiuta <win@wincent.com>