Accept tag names including numbers
This allows us to create useful tags such as "ruby2ruby"
and "wopen3".

Note that in making this change we replace the validation
regex defined in the "taggable" module, and make it consistent
with the one used in the tag model itself.

In fact, we actually fix a bug in the former version of the regex
(an unescaped literal period) which ironically allowed it to
accept space separated tags such as "foo bar baz" (as required)
but would also incorrectly accept strings like "foo_bar_baz".

This same bug also allowed blank strings to pass validation,
but on fixing the bug they no longer passed, and the should have.
So in addition we change the "allow_nil" exception to an
"allow_blank" one, to ensure that blank strings still pass.

The validation message for pending tag format ("may only contain
letters, numbers and periods") leaves a little to be desired,
but explicitly specifying the real requirements would be a little
cumbersome ("may only contain whitespace-separated tag names, each
of which may only contain words (numbers and letters) separated by
periods").

See:

  https://wincent.com/issues/1159

Signed-off-by: Wincent Colaiuta <win@wincent.com>