Redirect on successful tweet update in tweets#update
Previously we were calling "render" here, but that would
show possibly stale instance variable state. For example,
if we entered duplicate tags "foo bar foo", then "foo" and
"bar" taggings would be added to the record but in the
view we'd see the old value of the "pending tags" string,
"foo bar foo", which is illegal.

By doing a redirect instead, we see the state of the model
as it really is in the database.

Signed-off-by: Wincent Colaiuta <win@wincent.com>