Ruby syntax: add "statement" styling
The line between "keyword" and "statement" in the Ruby syntax rules is a
little blurry because words like "class" and "module" are keywords and
are colored blue, while words like "if" are statements and are colored
purple. Nevertheless both "class" (keyword) and "if" (statement) must be
paired with an "end" of the same color.

This is basically impossible using the simplistic algorithm implemented
here, so we are forced to basically use one color for both "keywords"
and "statements".

Nevertheless, there are some items which we want to use the purple
"statement" color, which don't have a corresponding "end" element. These
include thigns like "private", "protected" and "public".

Signed-off-by: Wincent Colaiuta <win@wincent.com>