Skip only anchor tags when highlighting
Given that we are syntax highlighting under controlled conditions (that
is, we know that all the text has already gone through the wikitext
translator, so we know that the only HTML tag that can be in the input
is the anchor tag) tweak the regexen to look for the anchor tag only.

We seek to avoid interleaved tags such as "<span><a></span></a>"
(invalid HTML) by treating the "<a> ... </a>" run as a single,
indivisible atom.

Signed-off-by: Wincent Colaiuta <win@wincent.com>