Handle hyperlinks inside string literals
Given input like:

  puts "Visit http://wincent.com/ for more info"

The wikitext translator will produce:

  puts "Visit <a href="http://wincent.com" class="external">http://wincent.com/</a> for more info"

The string-literal regex causes this to be broken up inappropriately as
follows:

  unstyled text:  puts<space>
  string literal: "Visit <a href="
  unstyled text:  http://wincent.com
  string literal " class="

And so on. This in turn leads to horribly corrupted HTML:

  puts <span class="string-literal-syntax">"Visit <a href="</span>http://wincent.com<span class="string-literal-syntax">" class="</span>

And so on. Note how the "href" target is now the totally unexpected:

  </span>http://wincent.com<span class=

One possible fix would be to add a check for this kind of unwanted
overlap between a non-skip (non-atom) match and a skip (atom) match, but
this in turn has some undesired side effects. For example, if we
suppress the highlighting of the string literal with the overlapping
anchor atom, we then find that the trailing quote here:

  ... for more info"

Is now identified as the start of a new string literal, and from that
point on the snippet is inappropriately highlighted.

The simplest solution, then, seems to be merely updating the
string-literal regex so that it knows about anchor tags and gobbles them
up as atoms.

Signed-off-by: Wincent Colaiuta <win@wincent.com>