Add "shell" syntax-highlighting
This is a very simple set of rules for highlight shell transcripts
(commands and their output); it is _not_ for highlighting actual Bash
scripts (for that I will be adding a "bash-syntax" rule set later on).

It handles two things: prompt/command lines (with optional line
continuations), and command output lines. It's difficult to go much
farther than that within the limitations of the highlighting algorithm.

Signed-off-by: Wincent Colaiuta <win@wincent.com>