Remove "private" CSS logic from posts#show view
The posts#show action uses comments retrieved using the "published"
method of the "Commentable" module. By definition, these will only ever
be comments that are both public and are not awaiting moderation.

This means that the logic here for applying the "private" class to
private comments is unneeded, as private comments can never appear in
the posts#show view.

Signed-off-by: Wincent Colaiuta <win@wincent.com>