Gracefully handle empty comments
Previously, if a user submitted an empty comment the comments controller
would perform the standard "render :action => 'new'". This naturally
didn't work, as the comments controller has neither a "new" action nor a
"new" template, and an exception was raised.

Now we gracefully handle this situation by redirecting to the "parent
path", which will vary according to whether the user is trying to post
a comment on a blog post, forum topic, issue, or (later) a wiki article.

Signed-off-by: Wincent Colaiuta <win@wincent.com>