Clarify per-post Atom feed title
Make title "wincent.com blog: foo bar" rather than "wincent.com:
comments on foo bar", to make it easier to distinguish between the
different kinds of feed.

In the near future will update the feed template to include the original
post, not just the comments, bringing it in line with the articles#show
Atom template. In this way users will be able to subscribe to per-post
feeds to check for updates to the original post.

Signed-off-by: Wincent Colaiuta <win@wincent.com>