Show comments in tweets#show
At the moment there is no UI for actually adding comments, only for
showing them. The reason is that the HTML format of tweets#show is
actually page cached, so we can't have a user-sensitive comments form in
place (that is, admins, logged-in users and anonymous visitors all see
different versions of the form); in a subsequent commit will add AJAX to
pull down the appropriate comments form on the fly.

Signed-off-by: Wincent Colaiuta <win@wincent.com>