Update tweets and comments sweepers for new commentable tweets
Now whenever a comment is added to a tweet we need to worry about
expiring the cache for both the HTML version of the tweet (already
implemented) and the Atom version of the tweet (soon to be implemented).

At the same time whenever a tweet is added, removed or updated we need
to make sure that not only the existing HTML version is expired, but
also the soon-to-be-implemented Atom version as well.

Signed-off-by: Wincent Colaiuta <win@wincent.com>